requirements.txt
streamlit
pandas
numpy
requests
google-genai